Reorged lens inits fixtures
