+++ pytest.ini
@@ -8,3 +8,4 @@ addopts =
     -p tests.fixtures.artifacts
     -p tests.fixtures.datasets
     -p tests.fixtures.frozen_tests
+    -p tests.fixtures.lens_inits
